add returnFieldsById prop to sources
