+++ components/airtable_oauth/actions/create-table/create-table.mjs
@@ -4,7 +4,7 @@ export default {
   key: "airtable_oauth-create-table",
   name: "Create Table",
   description: "Create a new table. [See the documentation](https://airtable.com/developers/web/api/create-table)",
-  version: "0.0.4",
+  version: "0.0.5",
   type: "action",
   props: {
     airtable,
