+++ components/airtable_oauth/actions/get-record-or-create/get-record-or-create.mjs
@@ -6,7 +6,7 @@ export default {
   key: "airtable_oauth-get-record-or-create",
   name: "Get Record Or Create",
   description: "Get a record from a table by record ID or create a new register.",
-  version: "0.0.5",
+  version: "0.0.6",
   type: "action",
   props: {
     ...common.props,
