+++ components/airtable_oauth/airtable_oauth.app.mjs
@@ -1,13 +1,19 @@
-import commonApp from "./common-app.mjs";
 import { ConfigurationError } from "@pipedream/platform";
 import { fieldTypeToPropType } from "./common/utils.mjs";
+import { axios } from "@pipedream/platform";
+import { SORT_DIRECTION_OPTIONS } from "./common/constants.mjs";
+import isEmpty from "lodash.isempty";
+import Bottleneck from "bottleneck";
+const limiter = new Bottleneck({
+  minTime: 250, // 4 requests per second
+  maxConcurrent: 1,
+});
+const axiosRateLimiter = limiter.wrap(axios);
 
 export default {
-  ...commonApp,
   type: "app",
   app: "airtable_oauth",
   propDefinitions: {
-    ...commonApp.propDefinitions,
     baseId: {
       type: "string",
       label: "Base",
@@ -190,13 +196,66 @@ export default {
       description: "An optional boolean value that lets you return field objects where the key is the field id. This defaults to `false`, which returns field objects where the key is the field name.",
       optional: true,
     },
+    sortDirection: {
+      type: "string",
+      label: "Sort: Direction",
+      description: "This field will be ignored if you don't select a field to sort by.",
+      options: SORT_DIRECTION_OPTIONS,
+      default: "desc",
+      optional: true,
+    },
+    maxRecords: {
+      type: "integer",
+      label: "Max Records",
+      description: "Optionally limit the maximum number of records to return. Leave blank to retrieve all records.",
+      optional: true,
+    },
+    filterByFormula: {
+      type: "string",
+      label: "Filter by Formula",
+      description: "Optionally provide a [formula](https://support.airtable.com/hc/en-us/articles/203255215-Formula-Field-Reference) used to filter records. The formula will be evaluated for each record, and if the result is not `0`, `false`, `\"\"`, `NaN`, `[]`, or `#Error!` the record will be included in the response. For example, to only include records where `Name` isn't empty, pass `NOT({Name} = '')`.",
+      optional: true,
+    },
+    records: {
+      type: "string",
+      label: "Records",
+      description: "Provide an array of objects. Each object should represent a new record with the column name as the key and the data to insert as the corresponding value (e.g., passing `[{\"foo\":\"bar\",\"id\":123},{\"foo\":\"baz\",\"id\":456}]` will create two records and with values added to the fields `foo` and `id`). The most common pattern is to reference an array of objects exported by a previous step (e.g., `{{steps.foo.$return_value}}`). You may also enter or construct a string that will `JSON.parse()` to an array of objects.",
+    },
+    typecast: {
+      type: "boolean",
+      label: "Typecast",
+      description: "The Airtable API will perform best-effort automatic data conversion from string values if the typecast parameter is `True`. Automatic conversion is disabled by default to ensure data integrity, but it may be helpful for integrating with 3rd party data sources.",
+      optional: true,
+    },
+    record: {
+      type: "object",
+      label: "Record",
+      description: "Enter the column name for the key and the corresponding column value. You can include all, some, or none of the field values. You may also pass a JSON object as a custom expression with key/value pairs representing columns and values (e.g., `{{ {\"foo\":\"bar\",\"id\":123} }}`). A common pattern is to reference an object exported by a previous step (e.g., `{{steps.foo.$return_value}}`).",
+    },
   },
   methods: {
-    ...commonApp.methods,
+    _baseUrl() {
+      return "https://api.airtable.com/v0";
+    },
     _headers() {
       return {
         Authorization: `Bearer ${this.$auth.oauth_access_token}`,
       };
+    },
+    async _makeRequest({
+      $ = this,
+      path,
+      rateLimited = true,
+      ...args
+    }) {
+      const config = {
+        url: `${this._baseUrl()}${path}`,
+        headers: this._headers(),
+        ...args,
+      };
+      return rateLimited
+        ? axiosRateLimiter($, config)
+        : axios($, config);
     },
     getRecord({
       baseId, tableId, recordId, ...args
@@ -316,6 +375,28 @@ export default {
         method: "DELETE",
         ...args,
       });
+    },
+    throwFormattedError(err) {
+      throw Error(`${err.error} - ${err.statusCode} - ${err.message}`);
+    },
+    validateRecord(record) {
+      if (typeof record !== "object") {
+        throw new Error("Airtable record isn't an object");
+      }
+      if (Array.isArray(record)) {
+        throw new Error("Airtable record is an array. Please pass an object, instead.");
+      }
+      if (isEmpty(record)) {
+        throw new Error("Airtable record data is empty");
+      }
+    },
+    validateRecordID(recordID) {
+      if (!recordID) {
+        throw new Error("Airtable record ID blank. Please pass a valid record ID");
+      }
+      if (!recordID.startsWith("rec")) {
+        throw new Error("Invalid Record ID. See documentation about Finding Airtable record IDs - https://support.airtable.com/docs/finding-airtable-record-ids.");
+      }
     },
   },
 };
