+++ components/airtable_oauth/sources/new-or-modified-field/new-or-modified-field.mjs
@@ -4,7 +4,7 @@ export default {
   name: "New or Modified Field",
   description: "Emit new event for each new or modified field in a table",
   key: "airtable_oauth-new-or-modified-field",
-  version: "0.0.4",
+  version: "0.0.5",
   type: "source",
   props: {
     ...common.props,
