+++ components/airtable_oauth/sources/new-or-modified-records-in-view/new-or-modified-records-in-view.mjs
@@ -1,13 +1,12 @@
 import base from "../common/common.mjs";
-import common from "./common.mjs";
+import moment from "moment";
 
 export default {
   ...base,
-  ...common,
   name: "New or Modified Records in View",
   description: "Emit new event for each new or modified record in a view",
   key: "airtable_oauth-new-or-modified-records-in-view",
-  version: "0.0.4",
+  version: "0.0.5",
   type: "source",
   props: {
     ...base.props,
@@ -34,5 +33,64 @@ export default {
       ],
       description: "The view ID to watch for changes.",
     },
+    returnFieldsByFieldId: {
+      propDefinition: [
+        base.props.airtable,
+        "returnFieldsByFieldId",
+      ],
+    },
+  },
+  async run(event) {
+    const {
+      baseId,
+      tableId,
+      viewId,
+    } = this;
+
+    const lastTimestamp = this._getLastTimestamp();
+    const params = {
+      view: viewId,
+      filterByFormula: `LAST_MODIFIED_TIME() > "${lastTimestamp}"`,
+      returnFieldsByFieldId: this.returnFieldsByFieldId,
+    };
+
+    const data = await this.airtable.listRecords({
+      baseId,
+      tableId,
+      params,
+    });
+
+    if (!data.records.length) {
+      console.log("No new or modified records.");
+      return;
+    }
+
+    const metadata = {
+      baseId,
+      tableId,
+      viewId,
+    };
+
+    let newRecords = 0, modifiedRecords = 0;
+    for (const record of data.records) {
+      if (!lastTimestamp || moment(record.createdTime) > moment(lastTimestamp)) {
+        record.type = "new_record";
+        newRecords++;
+      } else {
+        record.type = "record_modified";
+        modifiedRecords++;
+      }
+
+      record.metadata = metadata;
+
+      this.$emit(record, {
+        summary: `${record.type}: ${JSON.stringify(record.fields)}`,
+        id: record.id,
+      });
+    }
+    console.log(`Emitted ${newRecords} new records(s) and ${modifiedRecords} modified record(s).`);
+
+    // We keep track of the timestamp of the current invocation
+    this.updateLastTimestamp(event);
   },
 };
