+++ components/airtable_oauth/sources/new-or-modified-records/new-or-modified-records.mjs
@@ -7,7 +7,7 @@ export default {
   name: "New or Modified Records",
   key: "airtable_oauth-new-or-modified-records",
   description: "Emit new event for each new or modified record in a table",
-  version: "0.0.6",
+  version: "0.0.7",
   type: "source",
   props: {
     ...base.props,
@@ -37,6 +37,12 @@ export default {
       description: "Identifiers of spedific fields/columns to watch for updates",
       withLabel: true,
     },
+    returnFieldsByFieldId: {
+      propDefinition: [
+        base.props.airtable,
+        "returnFieldsByFieldId",
+      ],
+    },
   },
   methods: {
     ...base.methods,
@@ -47,22 +53,28 @@ export default {
       this.db.set("fieldValues", fieldValues);
     },
     updateFieldValues(newFieldValues, record) {
+      const fieldKey = this.returnFieldsByFieldId
+        ? "value"
+        : "label";
       for (const fieldId of this.fieldIds) {
         newFieldValues[record.id] = {
           ...newFieldValues[record.id],
-          [fieldId.value]: record.fields[fieldId.label] || null,
+          [fieldId.value]: record.fields[fieldId[fieldKey]] || null,
         };
       }
       return newFieldValues;
     },
     isUpdated(fieldValues, fieldIds, record) {
+      const fieldKey = this.returnFieldsByFieldId
+        ? "value"
+        : "label";
       for (const fieldId of fieldIds) {
-        if (!record.fields[fieldId.label]) {
-          record.fields[fieldId.label] = null;
+        if (!record.fields[fieldId[fieldKey]]) {
+          record.fields[fieldId[fieldKey]] = null;
         }
         if (fieldValues[record.id]
           && fieldValues[record.id][fieldId.value] !== undefined
-          && record.fields[fieldId.label] !== fieldValues[record.id][fieldId.value]
+          && record.fields[fieldId[fieldKey]] !== fieldValues[record.id][fieldId.value]
         ) {
           return true;
         }
@@ -80,13 +92,14 @@ export default {
     const lastTimestamp = this._getLastTimestamp();
     const fieldValues = this._getFieldValues();
     const isFirstRunWithFields = this.fieldIds && Object.keys(fieldValues).length === 0;
-    const params = isFirstRunWithFields
-      ? {}
-      : {
-        filterByFormula: `LAST_MODIFIED_TIME() > "${lastTimestamp}"`,
-      };
+    const params = {
+      returnFieldsByFieldId: this.returnFieldsByFieldId,
+    };
+    if (!isFirstRunWithFields) {
+      params.filterByFormula = `LAST_MODIFIED_TIME() > "${lastTimestamp}"`;
+    }
 
-    const data = await this.airtable.getRecords({
+    const data = await this.airtable.listRecords({
       baseId,
       tableId,
       params,
