+++ components/airtable_oauth/sources/new-records-in-view/new-records-in-view.mjs
@@ -1,13 +1,12 @@
 import base from "../common/common.mjs";
-import common from "./common.mjs";
+import moment from "moment";
 
 export default {
   ...base,
-  ...common,
   name: "New Records in View",
   description: "Emit new event for each new record in a view",
   key: "airtable_oauth-new-records-in-view",
-  version: "0.0.4",
+  version: "0.0.5",
   type: "source",
   props: {
     ...base.props,
@@ -34,5 +33,60 @@ export default {
       ],
       description: "The view ID to watch for changes.",
     },
+    returnFieldsByFieldId: {
+      propDefinition: [
+        base.props.airtable,
+        "returnFieldsByFieldId",
+      ],
+    },
+  },
+  async run() {
+    const {
+      baseId,
+      tableId,
+      viewId,
+    } = this;
+
+    const lastTimestamp = this._getLastTimestamp();
+    const params = {
+      view: viewId,
+      filterByFormula: `CREATED_TIME() > "${lastTimestamp}"`,
+      returnFieldsByFieldId: this.returnFieldsByFieldId,
+    };
+
+    const data = await this.airtable.liarRecords({
+      baseId,
+      tableId,
+      params,
+    });
+
+    if (!data.records.length) {
+      console.log("No new records.");
+      return;
+    }
+
+    const metadata = {
+      baseId,
+      tableId,
+      viewId,
+    };
+
+    let maxTimestamp;
+    let recordCount = 0;
+    for (const record of data.records) {
+      record.metadata = metadata;
+
+      this.$emit(record, {
+        ts: moment(record.createdTime).valueOf(),
+        summary: JSON.stringify(record.fields),
+        id: record.id,
+      });
+      if (!maxTimestamp || moment(record.createdTime).valueOf() > moment(maxTimestamp).valueOf()) {
+        maxTimestamp = record.createdTime;
+      }
+      recordCount++;
+    }
+    console.log(`Emitted ${recordCount} new records(s).`);
+    this._setLastTimestamp(maxTimestamp);
   },
 };
