--- components/airtable_oauth/sources/new-records/common.mjs
@@ -1,51 +1,0 @@
-import moment from "moment";
-
-export default {
-  async run() {
-    const {
-      baseId,
-      tableId,
-      viewId,
-    } = this;
-
-    const lastTimestamp = this._getLastTimestamp();
-    const params = {
-      filterByFormula: `CREATED_TIME() > "${lastTimestamp}"`,
-    };
-
-    const data = await this.airtable.getRecords({
-      baseId,
-      tableId,
-      params,
-    });
-
-    if (!data.records.length) {
-      console.log("No new records.");
-      return;
-    }
-
-    const metadata = {
-      baseId,
-      tableId,
-      viewId,
-    };
-
-    let maxTimestamp;
-    let recordCount = 0;
-    for (const record of data.records) {
-      record.metadata = metadata;
-
-      this.$emit(record, {
-        ts: moment(record.createdTime).valueOf(),
-        summary: JSON.stringify(record.fields),
-        id: record.id,
-      });
-      if (!maxTimestamp || moment(record.createdTime).valueOf() > moment(maxTimestamp).valueOf()) {
-        maxTimestamp = record.createdTime;
-      }
-      recordCount++;
-    }
-    console.log(`Emitted ${recordCount} new records(s).`);
-    this._setLastTimestamp(maxTimestamp);
-  },
-};
